Fix import of issues without body message
The CSV import threw an error when an issue didn't have a body defined.
Since it is a good practice to have a body defined, in the case of importing issues, we duplicate the title into the body. While this doesn't clarify the issue, it highlights the problem and still imports the issue.